separate plot files
